Fixing language issues and refactoring LanguageSelector into an external vue file
