+++ WebUI/src/assets/i18n/en-US.json
@@ -209,9 +209,18 @@
   "BACKEND_INFORMATION": "Information",
   "BACKEND_ENABLE": "Enable *",
   "BACKEND_ACTION": "Action",
+  "BACKEND_REQUIRED": "Required",
+  "BACKEND_OPTIONAL": "Optional",
   "BACKEND_REQUIRED_COMPONENTS_MESSAGE": "Before you can use the Intel AI Playground, we need to download some additional components. Please make sure you have a stable and unmetered internet connection.",
   "BACKEND_OPTIONAL_COMPONENTS_MESSAGE": "Optional components are not required for AI-Playground to work, but provide alternative functions. If you want to use them please click the info buttons to familiarize yourself with their terms and conditions before activating them.",
   "BACKEND_TERMS_AND_CONDITIONS": "* I have reviewed the optional component. I agree to all terms and conditions and would like to download and enable third-party software if applicable.",
+  "BACKEND_STATUS_RUNNING": "Running",
+  "BACKEND_STATUS_STOPPING": "Stopping",
+  "BACKEND_STATUS_STARTING": "Starting",
+  "BACKEND_STATUS_INSTALLED": "Installed",
+  "BACKEND_STATUS_NOT_INSTALLED": "Not Installed",
+  "BACKEND_STATUS_INSTALLING": "Installing",
+  "BACKEND_STATUS_FAILED": "Failed",
   "SETTINGS_IMAGE_MODE": "Mode",
   "SETTINGS_IMAGE_WORKFLOW": "Workflow",
   "WORKFLOW_RELOAD_INFO": "Reload workflows from disk.",
